+++ A Platformer Game by Richard and Tim/TestUI.h
@@ -71,7 +71,7 @@ struct TestSpaceWindow : ui::Window {
 					if (auto entity = ent.lock()){
 						vec2 p = viewtransform.getInverse().transformPoint(localMousePos());
 						vec2 disp = p - entity->getPosition();
-						double dist = abs(disp);
+						float dist = abs(disp);
 						if (dist > 0.0){
 							entity->velocity += disp * (float)(50.0 / pow(dist, 2));
 						}
